refactor(server): move collab runtime into package
