+++ packages/doc-server/package.json
@@ -6,7 +6,9 @@
   "description": "Proof SDK document, share, and collaboration server routes",
   "dependencies": {
     "@proof/core": "0.1.0",
-    "express": "^5.2.1"
+    "express": "^5.2.1",
+    "y-prosemirror": "^1.2.19",
+    "yjs": "^13.6.20"
   },
   "exports": {
     ".": "./src/index.ts",
